Support arbitrary byte sequence containers
Previously, the Command and Response types always used heapless::Vec to
store the payload.  This was awkward to use in non-embedded environments
where std::vec::Vec might be preferable, or in situations where a slice
is sufficient.

This patch relaxes this requirement and accepts any type implementing
AsRef<[u8]> (and TryFrom<&[u8]> for construction).