Fix Le parsing for simple APDUs with Lc > 0
In simple APDUs, Le is encoded by replacing 256 with 0.  In the Case 4S
from the spec, a simple APDU with Lc > 0, the command parser added one
to encoded length before decoding it.  This patch fixes this mistake and
adds a unit test for this case.  (Note that the other cases already
parse the Le correctly so no change is required there.)